updated train updater, added train price entity

+++ src/main/java/com/summerdev/travelstoragemanager/response/api/tutu/TutuRailwayCarriageResponse.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 public class TutuRailwayCarriageResponse {
     private String type;
-    private int price;
+    private double price;
 }
